+++ src/scripts/driver_script.js
@@ -1,11 +1,4 @@
-const fs = require('fs');
-const fetch = require('node-fetch');
 const { execSync } = require("child_process");
-
-// If one of these CSV files fails to download then some of the scripts using them will fail
-const CSV_FILE_NAMES = [ 'netfile_api_2018.csv', 'netfile_api_2019.csv', 'netfile_api_2020.csv' ];
-const ASSETS_PATH = `${__dirname}/../assets/data`;
-
 
 function processInput() {
 
@@ -24,29 +17,6 @@ function processInput() {
     .argv;
 
     return { downloadCSV: !args['skip-download'] }
-}
-
- /**
-  * This downloads files from Firebase Storage and saves them to the local system.
-  * @param {string[]} fileNames - names of the files to download from Firebase
-  * @param {string} localFilePath - path on the local system to save the downloaded files into
-  */
-async function downloadCSVFromFirebaseCloudStorage (fileNames, localFilePath){
-
-  const encodedPath = encodeURIComponent('data/');
-
-  for await (fileName of fileNames) {
-    const firebaseStorageLocation = 
-      `https://firebasestorage.googleapis.com/v0/b/san-diego-voters-voice.appspot.com/o/${encodedPath}${fileName}?alt=media`;
-
-    const fetchResponse = await fetch(firebaseStorageLocation);
-    const body = await fetchResponse.text();
-
-    fs.writeFileSync(`${localFilePath}/${fileName}`, body);
-
-    console.log(`Downloading remote file '${fileName}' from Firebase Storage \n To '${localFilePath}/${fileName}'`);
-  }
-
 }
 
 /**
@@ -77,14 +47,11 @@ function getPythonCommand() {
 (async () => {
   const input = processInput();
 
-  if (input.downloadCSV) {
-    await downloadCSVFromFirebaseCloudStorage(CSV_FILE_NAMES, ASSETS_PATH);
-  }
-
   const nodeCommand = 'node';
   const pythonCommand = getPythonCommand();
 
   if (!pythonCommand) { 
+
     console.log(`Error: Python version 3 not found`);
     process.exitCode = 1;
 
@@ -92,11 +59,18 @@ function getPythonCommand() {
   }
 
   /**
-   * 'calculation_download_gdrive_info.py' needs to be first since it updates/creates 
-   * the JSON files and sets committee name to use as key in later scripts.
+   * 'calculation_download_gdrive_info.py' needs to be run after preScripts since it 
+   * updates/creates the JSON files and sets committee name to use as key in later scripts.
    */
+  
+  let preScripts = [];
+
+  if (input.downloadCSV) {
+    preScripts.push( { command: nodeCommand, fileName: 'download_csv_from_firebase.js' } );
+  }
 
   const scripts = [
+    ... preScripts,
     { command: pythonCommand, fileName: 'calculation_download_gdrive_info.py' },
     { command: pythonCommand, fileName: 'candidate_calculation_amount_raised.py' },
     { command: pythonCommand, fileName: 'candidate_calculation_amount_spent.py' },
@@ -123,8 +97,8 @@ function getPythonCommand() {
     console.log(error.toString());
     console.log('Error: Candidate JSON files update NOT complete!');
     process.exitCode = 1;
-    return;
 
+    return;
   }
 
   console.log('Update of Candidate JSON files complete!');
